Update Device.cs
The handler function can be called for two reasons: 1) service discovery succeeded or 2) service discovery failed. In the later case `args.Error` will not be null and `_nativeDevice.Services` may be null. In that case, we cannot return from the handler without setting the result on the TaskCompletionSource. Doing so will certainly cause a deadlock because this task will never complete. The fact that the event handler is removed by `_nativeDevice.DiscoveredService -= handler;` ensures that. 

If an error occurs, the code will not complete with an empty list of services. Then the caller can do what they choose with that.