Fixed broken DeviceConnected / DeviceConnectionLost events
DeviceConnected / DeviceConnectionLost events now listen to the PropertyChanged event of the native device and now return the true connection status when it occurs.

I had to do an ugly hack to get a reference to the Device class to pass as an event parameter in the DeviceConnected and DeviceConnectionLost events. There really should be DeviceConnectionStatusChanged event instead of all these individual events and it should be in the Device class, not the Adapter class.