Add System.Threading.CancellationToken parameters to*NativeAsync methods.
Reason: peripheral:didWriteValueForCharacteristic:error: (CBPeripheral.WroteCharacteristicValue)
does not get called when connection is lost during a write
(perepheral moved out of range, shut off without sending disconnect notification, etc.).
This results in the Task created in TaskBuilder.FromEvent never transitioning to Completed or Failed,
causing the thread to hang.  Thus, a timeout should be implemented in the application to trigger a cancellation.
However, previously, the CancellationToken passed into the *Async methods did not get passed through to the *NativeAsync
methods, causing the cancellation to not end theTask and the Task to leak.