Added an option to ShowPowerAlert for iOS.
User can decide to enable or not the ShowPowerAlert on iOS. When this option is enable, a message will be displayed to the user to turn On the bluetooth in the Settings when in airplane mode. In our case, this message is annoying for our users.

Run the following within 'AppDelegateFinishedLaunching' before any other usage of the plugin: BleImplementation.ShowPowerAlert(false).

Default value of ShowPowerAlert is true. The ShowPowertAlert will be display even if there is not options provided to CBCentralManager constructor.