CLOUD-9959 do not close gatt untill BLE is disconnected
Closing the gatt while disconnecting BLE products will cause the Disconnect operationer to hang. This is due to the fact that the gatt is what propagates connection state events up which is waiting for the Disconnected event to happen. Closing the gatt after the disconnect is done will make the operation behave appropriately